+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - Deposit.account_type parameter
+- BrcodePayment.name parameter
 ### Fixed
 - Restrictive requests lib requirements